patches/openwrt: mpc85xx: support and use SPE float instructions
This is needed for Turris 1.x uboot compilation that is failing if soft float ABI is used.

As described in a subsequent commit inner message:

The usage of SPE[^3] ISA extension allows CPUs with PowerPC e500v2 cores (used in Turris 1.x routers, NXP P2020RDB
boards, ...) to perform native operations on single[^4] and double precision floats stored in GPRs[^5] as well
as vector operations using the whole 64-bit size of CPU registers (whose upper 32 bits are inaccessible otherwise).[^6]
This enables programs to be compiled for these CPUs without the need for floating-point arithmetic translation
leading to better performance.

[^3]: signal processing engine
[^4]: applies to e500v1 as well
[^5]: general purpose registers
[^6]: https://www.nxp.com/docs/en/reference-manual/SPEPEM.pdf