patches/luci: luci-mod-diagnostics: nslookup path: drop
We had the patch because we wanted to persuade the busybox to
install everything into /{s,}bin, but let's say it was not a good
solution for us. I verified that /usr/sbin/ntpd is now present in
correct path on router as it is in OpenWrt.

So, this patch is no longer required as the patch was removed.
See: https://gitlab.nic.cz/turris/os/build/-/commit/a4ce259c201c682a424161f418ceb4861a5140d4