patches/openwrt: Drop iflink property patch from Omnia kernel support
Drop the "net: add ndo for setting the iflink propert" from Omnia kernel
support.

This was needed on 5.4 for configuring CPU upstream ports for multi-CPU
DSA. This is going to be done differently in upstream kernel, and since
multi-CPU DSA is not backported for 5.15 yet, we don't need it now.