NEWS: 6.0.4: mention dropped packages for ISDN4Linux and other things
Packages related to old ISDN4Linux was dropped in Linux kernel since
version 5.3, but they were still present in OpenWrt. [1] In recent
commits, where it was updated our Linux 5.15 support the commit from
OpenWrt master branch is included, thus we need to let users know about
this breaking change, because it does not make sense to postpone it or
even have it in distribution anymore.

[1] https://github.com/openwrt/openwrt/commit/db55dea5fc047190af188f07018e99b0c7a4bdde