NEWS: 6.1.0: trim down changes as they are already mention
I dropped several things in Updated sections as they are already mention
in New Features, otherwise, the changelog is too long and users will not
get any clue from that. For example, if you said reForis update, then
they need to check the changelog, but the same thing is already in the
new features. Users don't need to know that we need to update backend
and Turris HW to get it working. The same applies for OpenVPN
credentials in reForis, which we added.