hostapd: Add support for card specific hacks
Hacky way to allow to override any settings for individual cards. This
makes it possible to fix configuration of cards that wrongly report
their capabilities.

Proper way is to somehow integrate it upfront, override reading of
possible capabilities and disable the options in all UIs and properly
report options that are not supported.

That would be quite some complex work and would take quite some time.
But this solution is simple enough and allows us to correct wrong
settings and make problematic cards work again.