patches/packages: syncthing: do not prevent from running
Enable syncthing in /etc/config by default to make sure when we
enable the service, syncthing can run without having to manually
enable it in the config.

Users, by default, expect that the service will run automatically and that
they will configure it while the service is running to see what was changed
and what is the behavior of each function there. This change is going to enable
syncthing to run after installing it instead of tweaking the config manually
(in LuCI or by hand). This is somehow different than how OpenWrt does it. They
force users to configure the app before running. Once they are confident that
the configuration looks good, they enable it, which is also good behavior.
Still, we need to improve the user experience and somehow make it different as
we want that even non-IT guys/girls can configure what they need and install
what they need in our simplified interface called (re)Foris, where we provided
list of packages, which they can install and if it does not run by default, it
looks like something is wrong. For syncthing, we also provide
Transmission WebApp, the integration for Turris OS, which you can find on the
landing page and easily click on it and start configuring it.