NEWS: Introduce version 6.2.4
Let's distinguish the version between HBK/HBT/HBS branches, otherwise,
it is same if it is not different and makes it harder for users to
know which branch or even the version they are using on the router.

This helps and also, it is good to mention that BIND was updated,
because it fixes three CVEs since I backported it to
OpenWrt 22.03 and OpenWrt 21.02.