patches/openwrt: static busybox: update config
The busybox package runs make oldconfig and sends a stream of enters,
this selects the default options for every option except for the ones
that expect a number. When the build systems gives it an enter, they
just prompt again which causes an infinite loop. Let's set them
explicitly to avoid them ever getting to oldconfig.