patches/openwrt/kernel: Rename turris-omnia-mcu compatible to turris-omnia-ctrl
Last year we prepared U-Boot so that when the turris-omnia-mcu driver is
finally in kernel, U-Boot will, before loading kernel, fixup Omnia's
device-tree by injecting reset-gpios to PCIe nodes (for board revisions
32+).

Now that the device-tree contains the node with "cznic,turris-omnia-mcu"
compatible, U-Boot does exactly that. But the problem is that the
pci-mvebu driver with reset-gpios available fails to enumerate the
ath10k card WLE900VX.

Rename the compatible of the driver to "cznic,turris-omnia-ctrl" so that
U-Boot does not inject the reset-gpios for now.