dnsmasq: Make log in ujail optional
Currently if you have ujail enabled and your logger hasn't started yet
when dnsmasq is being started, ujail fails to be created and dnsmasq
fails to start. So this patch makes the dependency on the logger
optional and allows dnsmasq with ujail to start even if there is no
logger daemon running.